requirements.txt
numpy
pandas
matplotlib
seaborn
plotly
scikit-learn
nltk
wordcloud
httpx
python-dotenv
tqdm
